+++ src/main/java/net/rogues/client/RoguesClient.java
@@ -1,18 +1,26 @@
 package net.rogues.client;
 
+import mod.azure.azurelibarmor.rewrite.render.armor.AzArmorRenderer;
+import mod.azure.azurelibarmor.rewrite.render.armor.AzArmorRendererRegistry;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.util.Identifier;
 import net.rogues.RoguesMod;
 import net.rogues.block.CustomBlocks;
+import net.rogues.client.armor.RogueArmorRenderer;
+import net.rogues.client.armor.WarriorArmorRenderer;
 import net.rogues.client.effect.ChargeParticles;
 import net.rogues.client.effect.DemoralizeParticles;
 import net.rogues.client.effect.ShatterParticles;
 import net.rogues.effect.Effects;
+import net.rogues.item.armor.Armors;
 import net.spell_engine.api.effect.CustomParticleStatusEffect;
+import net.spell_engine.api.item.armor.Armor;
 import net.spell_engine.api.render.StunParticleSpawner;
 import net.spell_engine.client.gui.SpellTooltip;
+
+import java.util.function.Supplier;
 
 public class RoguesClient implements ClientModInitializer {
     @Override
@@ -42,5 +50,16 @@ public class RoguesClient implements ClientModInitializer {
             description = description.replace(SpellTooltip.placeholder("damage_reduction"), damageReduction);
             return description;
         });
+
+        registerArmorRenderer(Armors.RogueArmorSet_t1, RogueArmorRenderer::rogue);
+        registerArmorRenderer(Armors.RogueArmorSet_t2, RogueArmorRenderer::assassin);
+        registerArmorRenderer(Armors.RogueArmorSet_t3, RogueArmorRenderer::netheriteAssassin);
+        registerArmorRenderer(Armors.WarriorArmorSet_t1, WarriorArmorRenderer::warrior);
+        registerArmorRenderer(Armors.WarriorArmorSet_t2, WarriorArmorRenderer::berserker);
+        registerArmorRenderer(Armors.WarriorArmorSet_t3, WarriorArmorRenderer::netheriteBerserker);
+    }
+
+    private static void registerArmorRenderer(Armor.Set set, Supplier<AzArmorRenderer> armorRendererSupplier) {
+        AzArmorRendererRegistry.register(armorRendererSupplier, set.head, set.chest, set.legs, set.feet);
     }
 }