+++ src/main/java/net/rogues/client/armor/RogueArmorRenderer.java
@@ -1,10 +1,25 @@
 package net.rogues.client.armor;
 
-import mod.azure.azurelibarmor.common.api.client.renderer.GeoArmorRenderer;
-import net.rogues.item.armor.RogueArmor;
+import mod.azure.azurelibarmor.rewrite.render.armor.AzArmorRenderer;
+import mod.azure.azurelibarmor.rewrite.render.armor.AzArmorRendererConfig;
+import net.minecraft.util.Identifier;
+import net.rogues.RoguesMod;
 
-public class RogueArmorRenderer extends GeoArmorRenderer<RogueArmor> {
-    public RogueArmorRenderer() {
-        super(new RogueArmorModel());
+public class RogueArmorRenderer extends AzArmorRenderer {
+    public static RogueArmorRenderer rogue() {
+        return new RogueArmorRenderer("rogue_armor", "rogue_armor");
+    }
+    public static RogueArmorRenderer assassin() {
+        return new RogueArmorRenderer("rogue_armor", "assassin_armor");
+    }
+    public static RogueArmorRenderer netheriteAssassin() {
+        return new RogueArmorRenderer("rogue_armor", "netherite_assassin_armor");
+    }
+
+    public RogueArmorRenderer(String modelName, String textureName) {
+        super(AzArmorRendererConfig.builder(
+                Identifier.of(RoguesMod.NAMESPACE, "geo/" + modelName + ".geo.json"),
+                Identifier.of(RoguesMod.NAMESPACE, "textures/armor/" + textureName + ".png")
+        ).build());
     }
 }